targeting: Preserve include paths of header files
Flat include directory based header files cannot be used
as the current library directory structure does not allow
grouping of all header files in a single include directory.

So the file paths need to be preserved in the included directory
to avoid transitive files included to not throw file not found error.

Signed-off-by: Gopichand Paturi <gopichandpaturi@gmail.com>